chore(resources): Mover rems.py a resources
